spinner item calculates given type

+++ UnitCoverter/app/src/main/java/com/example/unitcoverter/Converter.java
@@ -1,10 +1,10 @@
 package com.example.unitcoverter;
 
 public class Converter {
-    public static float toCelcius(float fah) {
+    public static double toCelcius(double fah) {
         return (fah - 32) * 5/9;
     }
-    public static float toKilograms(float pounds){ return pounds * (float) 0.45359237; }
-    public static float toFeet(float inches){ return inches/12; }
-    public static float toMl(float ounces) { return ounces * (float) 29.57353; }
+    public static double toKilograms(double pounds){ return pounds * (float) 0.45359237; }
+    public static double toFeet(double inches){ return inches/12; }
+    public static double toMl(double ounces) { return ounces * (float) 29.57353; }
 }
